backend/requirements.txt
fastapi==0.134.0
uvicorn==0.41.0
pydantic==2.12.5
google-genai==1.65.0
pillow==12.1.1